+++ A3A/addons/gear/ace/CfgFunctions.hpp
@@ -1,0 +1,12 @@
+
+class CfgFunctions
+{
+    class A3A
+    {
+        class medical {
+            file = "x\A3A\addons\gear\ace\functions";
+            class coffeeIV {};
+            class drinkMed {};
+        };
+    };
+};
